Completed refactor
Refactored server components through removal of `router.refresh()` and other such functions.

This involved adding `client.tsx` files to the `(standalone)` and `(dashboard)` areas of the code for areas such as `[taskId]->page.tsx`

This allows the `page.tsx` files in `[taskId]`, `[projectId]`, etc. to use their respective `client.tsx` to pull the information necessary for loading the page efficiently and cleanly.