Theme Switcher Completed
Light and dark mode are implemented to a somewhat full degree. 95% there, just needs some polishing here and there. No custom themes yet, but it is in the works. The form is there, just disabled and whatnot, but the update function isn't working properly or as expected. Hopefully can get that rolling out soon